refactor: 拆分 GeneratePage 中路设置面板为 GenerationSettingsPanel 独立组件
